Add music directory parsing and server functionality
This commit introduces MP3 directory parsing functionality to get tag info and store it in H2 database. It also sets up a simple Jetty HTTP server to serve a webpage displaying the parsed song info. For the same, the respective necessary classes are created. A .gitignore file has been added to ignore unnecessary files or directories.