Update dependencies and refactor resourceBase setting in Main
The commit includes changes in Main.java where the usage of ServletContextHandler has been updated from the jetty package to the jetty.ee10 package. Also, a method call for setting the resource base in the servlet context has been updated: the setResourceBase method was changed to setBaseResourceAsString. Apart from that, pom.xml was updated to include additional project dependencies and pluginManagement section with several new plugins.